python-snake-2/requirements.txt
Flask==3.0.2
Flask-CORS==4.0.0
pymongo==4.6.2
python-dotenv==1.0.1
pytest==8.0.2 